+++ packages/expo-router/src/layouts/withLayoutContext.tsx
@@ -164,7 +164,13 @@ export function withLayoutContext<
       return (
         <IsWithinLayoutContext value>
           <GuardContextProvider node={node} guardedRedirects={guardedRedirects}>
-            <Nav {...props} id={contextKey} ref={ref} children={sorted} />
+            <Nav
+              {...props}
+              id={contextKey}
+              ref={ref}
+              children={sorted}
+              UNSTABLE_routeNamesAreStatic
+            />
           </GuardContextProvider>
         </IsWithinLayoutContext>
       );